Fix Bug creating group calendars over and over
When running the script for creating default group calendars,
the script should check if there is already a default group calendar, by
checking a tag in the .Radicale.props of each collection. Before the fix
it would only create a calendar, if there is only default calendars and
no custom calendars present.